✨ Add `listAllValuesOnFocus` prop.
