+++ README.md
@@ -24,7 +24,9 @@
 <a><img alt="GitHub Repo stars" src="https://img.shields.io/github/stars/anubra266/choc-autocomplete?logo=github&style=for-the-badge">
 
 <!-- ALL-CONTRIBUTORS-BADGE:START - Do not remove or modify this section -->
+
 [![All Contributors](https://img.shields.io/badge/all_contributors-3-orange.svg?style=flat-square)](#contributors-)
+
 <!-- ALL-CONTRIBUTORS-BADGE:END -->
 
 </a>
@@ -483,6 +485,12 @@ boolean | MaybeRenderProp<{ value: Item["value"] }>
             <td>isReadOnly</td>
             <td>boolean</td>
             <td>Make the component read-only</td>
+            <td>false</td>
+        </tr>
+         <tr>
+            <td>listAllValuesOnFocus</td>
+            <td>boolean</td>
+            <td>Show all suggestions when user focuses the input, while it's not empty.</td>
             <td>false</td>
         </tr>
         <tr>
